Add title to StripeDialog (#803)
While this is hidden in our stylesheet, it can show up in other stylesheets. The null default is the app ID, which is ugly, and an empty string can look really strange. I figure setting something sane like this is fine for when it does show up but isn't critical info for us at all since we're hiding it.