Rewrite ScreenshotCache (#1350)
When the screenshot cache folder goes over 100MB, there's some code in this class that gets stuck in an infinite loop, causing AppCenter to use 100% of a CPU core. And starts using another core every time you open a new app info page. Currently the only way to stop it is to kill appcenter and delete the screenshot cache folder. I've marked this as high priority as a result.

When looking into that issue I found the code here didn't really fit with our code style, so I did the following:

    Added a copyright header
    Removed yoda conditions
    Used simpler GLib methods rather than doing stuff manually with C stdlib and Soup
    Removed while loops that had the potential to go on forever if stuff went wrong
    Returned booleans to indicate success or not instead of integers
    Removed the inline thread and just used async methods instead because we already use a thread to call on this class

By the time I was finished doing that, it was basically rewritten.

It's easier to review the code in split diff view, or just read the new file.