Put the release notes description label inside a Gtk.Grid
This is the same issue that we had for issues, but now for the description label itself. It seems that in a Gtk.Box with wrapped labels will not do its height correctly, if there are grids abound.

Putting the label inside a Gtk.Grid will make it also respect the natural height and will fix another set of problems found in other description labels.