Various fixes in AppListUpdateView (#2237)
* Set vexpand property to false to avoid extra unnecesary space in listbox.

* Disable refresh_menuitem when clicking it an renabling it when job is complete

This is done to avoid multiple update check commands that could create duplicated update element in the list.

* Remove updated and installed list along with labels when checking for updates.

This commit also toggles the updates message properly when trying to check new updates. Even when the app closes. This removes all installed and updated apps so iit clears the view whn tyring to update again.

* Removing the placeholder methods.

Sionce this only appear when lists are empty, there not need to explicitly remove  since we arealready removing list data.

* Fixed a release notes icon bug when installing/cancelling and update

The release notes icon would appear when canelling an update of an application that doew nost have a description (and thus the icon should not appear)

* Fix for #529 Appcenter sometimes opens the updates view.

This also fix an issue where installed apps would not show up if your did not open the AppListUpdateView until the update worker ended.

* Use the refrsh action directlt action to update the refresh_menuitem button instead of action_name.

This helps to use the shortuct  to update elements and makes the sentiive property of the button more consistent when it starts and ends.

* Move the clearing of the update_liststore to a better place.

Apartfor being a better palce organizational asking. Thsi also avoids a bug when if you the application while listing updates, it will show a bad update list badge and also would keep the incomplete update list when re-opening the app.

* Restore the placeholder setup.

This is needed for when there are no updates. oops.

* Check the vexpand value of list_box after adding the possible items in installed_liststore

---------

Co-authored-by: italo-capasso <italo@glitchypixel.com.co>