Emit on_metadata_remote_preprocessed always from the main thread (#2306)
When we refresh the cache we run on the worker thread and would emit this signal from there.

We connect to it from the appicon though and never go to the main thread so we change gtk properties etc. from the wrong thread which would lead to crashes, sometimes a only white window, app entries dissapearing etc.