Move action cancellable to ChangeInfo
Instead of manually passing the cancellable, include it in change info.
This makes it easier to keep the change info status in sync with the
cancelled state. Notably this allows to fix an issue where clicking
cancel seemingly didn't do anything until the backend actually drops out
of the transaction.