Reload installed applications synchronously
There are only two occasions where we have to do this and in both cases
it's ok to do it synchronously:
- Startup: We have to wait anyways
- External change to the apps: Then you're probably not in the AppCenter
anyways.

Also fetching installed applications is really fast so it wouldn't
even be noticable.

Also this will safe us the is_package_installed call which does
exactly the same IO as getting all installed apps and is already
happening synchronously multiple times on startup and when using
AppCenter e.g. when installing apps.